Create new database structure which supports multiple plays and actor read positions.
A new table, "plays", keeps track of the different plays supported. It will no longer be necessary to have separate tables for each play.

Actors now store their name as one column, "name", instead of two, "first_name" and "last_name".

Characters now are associated with plays via a foreign key.

Where appropriate, foreign keys now have ON DELETE CASCADE to ensure that related data is deleted if the entry it refers to is deleted.
For example, deleting a play will delete all the characters and read positions that refer to that play.
One exception is for character and actor associations. If we delete an actor, we do not want to delete a character but instead want to mark it as available for assignment to another actor (represented by NULL).